slabs-forecasting/requirements.txt
numpy==1.19.*
pandas==1.2.*
scikit-learn==0.24.1
scipy==1.4.*
SQLAlchemy==1.3.*
PyMySQL==1.0.2
setuptools<50.0.0
statsmodels==0.11.*
pmdarima==1.7.*
jsonschema==3.2.*
pyarrow==1.0.*
tensorflow==2.4.1
keras-tuner==1.0.*
json-tricks==3.15.*
lightgbm==3.1.1
google-auth==1.24.0
google-cloud-bigquery==2.6.*
google-cloud-storage==1.35.*
google-cloud-logging==2.0.2
google-cloud-ndb==1.11.1
pybigquery==0.5.0
pandas-gbq==0.14.1
python-dateutil==2.8.*
zstandard==0.15.2
tenacity==7.0.0
crcmod==1.7
